# DAR/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api rest on a task application",
    "version": "1.0.0",
    "title": "DAR: APP APP HZ"
  },
  "host": "api-manag-tasks.azure-api.net",
  "basePath": "/v1",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/tasks/listAll": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Finds All tasks",
        "description": "Find all tasks in DB",
        "operationId": "ShowAllTasks",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              }
            }
          },
          "404": {
            "description": "Null data"
          }
        },
        "security": [
          {
            "tasks_auth": [
              "write:tasks",
              "read:tasks"
            ]
          }
        ]
      }
    },
    "/tasks/findById/{id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Find task by ID",
        "description": "Returns a single task",
        "operationId": "findById",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of task to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/tasks/newTask": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Add new task",
        "description": "this function is only used to add new tasks ",
        "operationId": "AddNewTask",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created task object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/tasks/update": {
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Update an existing task",
        "description": "",
        "operationId": "modify",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Task object that exist in DB",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Task"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Task not found"
          },
          "405": {
            "description": "Validation exception"
          }
        },
        "security": [
          {
            "tasks_auth": [
              "write:tasks",
              "read:tasks"
            ]
          }
        ]
      }
    },
    "/tasks/deleteById": {
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Deletes a task by its id",
        "operationId": "deleteById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Id of the task",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "tasks_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:tasks": "modify tasks",
        "read:tasks": "read your tasks"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Task": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "description": {
          "type": "string",
          "example": "description 1"
        },
        "status": {
          "type": "string",
          "description": "task status",
          "enum": [
            "TO-DO",
            "Work in progess",
            "DONE"
          ]
        }
      },
      "xml": {
        "name": "Tasks"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}